+++ CHANGES.md
@@ -1,9 +1,9 @@
 #5.0.20 2017-??-??
+
+#5.0.19 2017-05-09
  - 2017-05-04 Updated translations, thanks to all translators.
  - 2017-05-03 Fixed bug#[12803](https://bugs.otrs.org/show_bug.cgi?id=12803) - TranslationsUpdate console command does not consider strings from var/packagesetup folder.
  - 2017-05-03 Added a web timeout config option for the support data collection and suppress the log message, if the internal web request for the support data collection doesn't work.
-
-#5.0.19 2017-05-09
  - 2017-05-02 Updated translations, thanks to all translators.
  - 2017-05-02 Improved error message, when trying to install an incompatible version of OTRS Business Solution™.
  - 2017-04-28 Fixed bug#[10556](https://bugs.otrs.org/show_bug.cgi?id=10556)(PR#1737) - Missing library used in IMAP authentication mechanism. Thanks to S7!